refactoring

+++ default_greek/uninstall.php
@@ -12,15 +12,14 @@ $query = $this->db->query(
     FROM " . $this->db->table("languages") . " 
 	WHERE code='" . $lngCode . "' AND directory='" . $lngDir . "'");
 $language_id = $query->row["language_id"];
+if($language_id) {
 //delete menu
-$storefront_menu = new AMenu_Storefront();
-$storefront_menu->deleteLanguage($language_id);
+    $storefront_menu = new AMenu_Storefront();
+    $storefront_menu->deleteLanguage($language_id);
 
 //delete all other language related tables
-$lm = new ALanguageManager($this->registry, $lngCode);
-$lm->deleteAllLanguageEntries($language_id);
-
-//delete language
-$this->db->query("DELETE FROM " . $this->db->table("languages") . " WHERE `code`='" . $lngCode . "'");
+    $lm = new ALanguageManager($this->registry, $lngCode);
+    $lm->deleteAllLanguageEntries($language_id);
+}
 
 $this->cache->remove("localization");
