+++ default_russian/uninstall.php
@@ -31,15 +31,16 @@ $query = $this->db->query(
 );
 
 $language_id = (int)$query->row['language_id'];
+if($language_id) {
 //delete menu
-$storefront_menu = new AMenu_Storefront();
-$storefront_menu->deleteLanguage($language_id);
+    $storefront_menu = new AMenu_Storefront();
+    $storefront_menu->deleteLanguage($language_id);
 
 //delete all other language related tables
-$lm = new ALanguageManager($this->registry, $language_code);
-$lm->deleteAllLanguageEntries($language_id);
-
+    $lm = new ALanguageManager($this->registry, $language_code);
+    $lm->deleteAllLanguageEntries($language_id);
+}
 //delete language
 $this->db->query("DELETE FROM " . $this->db->table("languages") . " WHERE `code`='" . $language_code . "'");
 
-$this->cache->remove('localization');
+$this->cache->remove('*');
